Porcentaje comision en empresa cambiado de int a float

+++ Modelo/Empresa.cs
@@ -18,7 +18,7 @@ namespace PagoAgilFrba.Modelo
         private String direccion;
         private int idRubro;
         private Boolean activo;
-        private int porcentajeComision;
+        private double porcentajeComision;
 
         private IList<SqlParameter> parametros = new List<SqlParameter>();
         private DBMapper mapper = new DBMapper();
@@ -39,7 +39,7 @@ namespace PagoAgilFrba.Modelo
         { this.direccion = direccion; }
         public void setIDRubro(int idRubro)
         { this.idRubro = idRubro; }
-        public void setPorcentajeComision(int porcentajeComision)
+        public void setPorcentajeComision(double porcentajeComision)
         {
             if (Util.EstaEntre0y100(porcentajeComision))
                 this.porcentajeComision = porcentajeComision;
@@ -52,7 +52,7 @@ namespace PagoAgilFrba.Modelo
         public String getCuit()             { return this.cuit; }
         public String getDireccion()        { return this.direccion; }
         public int getIDRubro()             { return this.idRubro; }
-        public int getPorcentajeComision()  { return this.porcentajeComision; }
+        public double getPorcentajeComision()  { return this.porcentajeComision; }
 
 
 
@@ -91,7 +91,7 @@ namespace PagoAgilFrba.Modelo
             this.direccion = Convert.ToString(reader["emp_direccion"]);
             this.idRubro = Convert.ToInt32(reader["id_rubro"]);
             this.activo = Convert.ToBoolean(reader["estado_habilitacion"]);
-            this.porcentajeComision = Convert.ToInt32(reader["porcentaje_comision"]);
+            this.porcentajeComision = Convert.ToDouble(reader["porcentaje_comision"]);
         }
 
         #endregion
